Rename lp_discr->discr_space and reorder __init__.py
