+++ odl/discr/__init__.py
@@ -10,21 +10,17 @@
 
 from __future__ import absolute_import
 
+from . import discr_utils
+from .diff_ops import *
+from .discr_ops import *
+from .discr_space import *
+from .grid import *
+from .partition import *
+
 __all__ = ()
 
-from .grid import *
 __all__ += grid.__all__
-
-from .partition import *
 __all__ += partition.__all__
-
-from .lp_discr import *
-__all__ += lp_discr.__all__
-
-from .discr_ops import *
+__all__ += discr_space.__all__
 __all__ += discr_ops.__all__
-
-from .diff_ops import *
 __all__ += diff_ops.__all__
-
-from . import discr_utils
